+++ NEWS.md
@@ -1,3 +1,8 @@
+
+# narampeg 0.1.1  2024-08-11
+
+* Decode multiple audio frames in a single call
+
 # narampeg 0.1.0  2024-08-10
 
 * Initial release